Refactor admin route to be RestFul
