+++ client/src/api/api-paths.js
@@ -25,14 +25,11 @@ const adminApiPaths = (isBuildWithAll || isBuildWithAdmin) && {
   exportPlacesExamsStatsKpi: `${apiPrefix}/admin/stats-places-exams`,
   login: `${apiPrefix}/auth/admin/token`,
   myProfile: `${apiPrefix}/admin/me`,
-  place: `${apiPrefix}/admin/place`,
   places: `${apiPrefix}/admin/places`,
-  reservations: `${apiPrefix}/admin/reservations`,
   uploadCandidatsJson: `${apiPrefix}/admin/candidats`,
   searchCandidats: `${apiPrefix}/admin/candidats?matching=`,
   searchWhitelisted: `${apiPrefix}/admin/whitelisted?matching=`,
   inspecteurs: `${apiPrefix}/admin/inspecteurs`,
-  uploadPlacesCSV: `${apiPrefix}/admin/places`,
   verifyToken: `${apiPrefix}/auth/admin/verify-token`,
   whitelist: `${apiPrefix}/admin/whitelisted`,
   generateBordereaux: `${apiPrefix}/admin/bordereaux`,
