+++ server/src/routes/admin/reservations.controllers.js
@@ -1,89 +1,91 @@
-import { findPlaceByIdAndPopulate } from '../../models/place'
-import { findUserById } from '../../models/user'
-import { appLogger } from '../../util'
-import {
-  PLACE_IS_NOT_BOOKED,
-  UNKNOWN_ERROR_REMOVE_RESA,
-} from './message.constants'
-import { removeReservationPlaceByAdmin } from './places-business'
+// // TODO: Remove this file
 
-export const removeReservationByAdmin = async (req, res) => {
-  const id = req.params.id
+// import { findPlaceByIdAndPopulate } from '../../models/place'
+// import { findUserById } from '../../models/user'
+// import { appLogger } from '../../util'
+// import {
+//   PLACE_IS_NOT_BOOKED,
+//   UNKNOWN_ERROR_REMOVE_RESA,
+// } from './message.constants'
+// import { removeReservationPlaceByAdmin } from './places.business'
 
-  const loggerContent = {
-    section: 'admin-delete-resa',
-    admin: req.userId,
-    placeId: id,
-  }
-  const admin = await findUserById(req.userId)
+// export const removeReservationByAdmin = async (req, res) => {
+//   const id = req.params.id
 
-  if (!admin) {
-    const message = 'Utilisateur non trouvé'
-    appLogger.warn({ ...loggerContent, description: message })
-    return res.status(404).send({
-      success: false,
-      message,
-    })
-  }
+//   const loggerContent = {
+//     section: 'admin-delete-resa',
+//     admin: req.userId,
+//     place: id,
+//   }
+//   const admin = await findUserById(req.userId)
 
-  if (!id) {
-    const message = 'Place non trouvée'
-    appLogger.warn({ ...loggerContent, description: message })
-    return res.status(404).send({
-      success: false,
-      message,
-    })
-  }
+//   if (!admin) {
+//     const message = 'Utilisateur non trouvé'
+//     appLogger.warn({ ...loggerContent, description: message })
+//     return res.status(404).send({
+//       success: false,
+//       message,
+//     })
+//   }
 
-  loggerContent.action = 'DELETE_RESA_BY_ADMIN'
+//   if (!id) {
+//     const message = 'Place non trouvée'
+//     appLogger.warn({ ...loggerContent, description: message })
+//     return res.status(404).send({
+//       success: false,
+//       message,
+//     })
+//   }
 
-  appLogger.info({
-    ...loggerContent,
-    description: 'Supprimer la réservation candidat',
-  })
+//   loggerContent.action = 'DELETE_RESA_BY_ADMIN'
 
-  // Have a reservation
-  const place = await findPlaceByIdAndPopulate(id, { candidat: true })
-  if (!place) {
-    loggerContent.action = 'FIND_PLACE'
-    const message = 'Place non trouvée'
-    appLogger.warn({ ...loggerContent, description: message })
-    return res.status(404).send({
-      success: false,
-      message,
-    })
-  }
+//   appLogger.info({
+//     ...loggerContent,
+//     description: 'Supprimer la réservation candidat',
+//   })
 
-  const { candidat } = place
-  if (!candidat) {
-    loggerContent.action = 'NOT_RESA'
-    const message = PLACE_IS_NOT_BOOKED
-    appLogger.warn({
-      ...loggerContent,
-      description: message,
-    })
+//   // Have a reservation
+//   const place = await findPlaceByIdAndPopulate(id, { candidat: true })
+//   if (!place) {
+//     loggerContent.action = 'FIND_PLACE'
+//     const message = 'Place non trouvée'
+//     appLogger.warn({ ...loggerContent, description: message })
+//     return res.status(404).send({
+//       success: false,
+//       message,
+//     })
+//   }
 
-    return res.status(400).send({
-      success: false,
-      message,
-    })
-  }
-  try {
-    // Annulation reservation
-    const result = await removeReservationPlaceByAdmin(place, candidat, admin)
+//   const { candidat } = place
+//   if (!candidat) {
+//     loggerContent.action = 'NOT_RESA'
+//     const message = PLACE_IS_NOT_BOOKED
+//     appLogger.warn({
+//       ...loggerContent,
+//       description: message,
+//     })
 
-    appLogger.info({ ...loggerContent, result })
-    return res.status(200).json({ success: true, ...result })
-  } catch (error) {
-    appLogger.error({
-      ...loggerContent,
-      action: loggerContent.action || 'ERROR',
-      description: error.message,
-      error,
-    })
-    return res.status(500).send({
-      success: false,
-      message: UNKNOWN_ERROR_REMOVE_RESA,
-    })
-  }
-}
+//     return res.status(400).send({
+//       success: false,
+//       message,
+//     })
+//   }
+//   try {
+//     // Annulation reservation
+//     const result = await removeReservationPlaceByAdmin(place, candidat, admin)
+
+//     appLogger.info({ ...loggerContent, result })
+//     return res.status(200).json({ success: true, ...result })
+//   } catch (error) {
+//     appLogger.error({
+//       ...loggerContent,
+//       action: loggerContent.action || 'ERROR',
+//       description: error.message,
+//       error,
+//     })
+//     return res.status(500).send({
+//       success: false,
+//       message: UNKNOWN_ERROR_REMOVE_RESA,
+//     })
+//   }
+// }
